+++ Pandora/Paradigm/Structure/Ability.hs
@@ -6,4 +6,3 @@ import Pandora.Paradigm.Structure.Ability.Substructure as Exports
 import Pandora.Paradigm.Structure.Ability.Morphable as Exports
 import Pandora.Paradigm.Structure.Ability.Possible as Exports
 import Pandora.Paradigm.Structure.Ability.Accessible as Exports
-import Pandora.Paradigm.Structure.Ability.Nonempty as Exports