+++ Pandora/Paradigm/Structure/Modification.hs
@@ -3,5 +3,6 @@ module Pandora.Paradigm.Structure.Modification (module Exports) where
 import Pandora.Paradigm.Structure.Modification.Turnover as Exports
 import Pandora.Paradigm.Structure.Modification.Tape as Exports
 import Pandora.Paradigm.Structure.Modification.Prefixed as Exports
+import Pandora.Paradigm.Structure.Modification.Nonempty as Exports
 import Pandora.Paradigm.Structure.Modification.Comprehension as Exports
 import Pandora.Paradigm.Structure.Modification.Combinative as Exports