+++ Pandora/Paradigm/Structure/Modification/Nonempty.hs
@@ -1,4 +1,4 @@
-module Pandora.Paradigm.Structure.Ability.Nonempty where
+module Pandora.Paradigm.Structure.Modification.Nonempty where
 
 -- | Type synonymous for at least one element data structure
 type family Nonempty (s :: * -> *) = (r :: * -> *) | r -> s